Feat : 서버 다운 관련 리다이렉팅 추가
