+++ app/(beforeLogin)/layout.tsx
@@ -1,5 +1,6 @@
 import { ReactNode } from 'react';
-import RedirectDashboard from '@/app/_components/common/RedirectDashboard';
+
+import LandingRedirectHandler from '@/app/_components/common/LandingRedirectHandler';
 
 type Props = { children: ReactNode };
 
@@ -7,7 +8,7 @@ const BeforeLoginLayout = ({ children }: Props) => {
   return (
     <>
       {children}
-      <RedirectDashboard />
+      <LandingRedirectHandler />
     </>
   );
 };
