+++ app/_constants/api.ts
@@ -8,6 +8,7 @@ export const END_POINTS = {
   STUDY_ENTER: '/study/enter-study',
   USER_INFO: '/member/info',
   UPDATE_CODE: '/member/edit-code',
+  HEALTH: '/actuator/health',
   DELETE_USER: (memberId: string) => `/member/${memberId}/info`,
   PROFILE_UPLOAD: (memberId: string) => `/member/${memberId}/upload`,
   RECENT_LOGS: (studyId: string) => `/study/${studyId}/logs`,
